requirements.txt
streamlit==1.33.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
PyJWT==2.8.0
